Refactor project structure to use pyproject.toml
This commit represents a significant refactor of the project structure. It replaces the setup.py and requirements.txt with a pyproject.toml which contains all the necessary information about the project and its dependencies. The python version in Github workflow is also updated from "3.10" to "3.12". Changes also include modifications to test and linting configurations for better compatibility with the refactored structure.

Signed-off-by: Steven K <steven@pointcircle.com>